+++ awl/tests/urls.py
@@ -7,4 +7,7 @@ urlpatterns = [
     url(r'admin/', include(admin.site.urls)),
 
     url(r'rankedmodel/', include(ranked_urls)),
+
+    url(r'awl_test_views/test_view_for_messages/$', 
+        'awl.tests.views.test_view_for_messages', ),
 ]